requirements.txt
flask==2.3.3
flask-cors==4.0.0
flask-sqlalchemy==3.1.1
sqlalchemy==2.0.23
python-dotenv==1.0.0
openpyxl==3.1.2
pandas==2.1.1